+++ Changelog.md
@@ -3,6 +3,7 @@
 ### v3.1.0
 
 - Added `firstCapitalized` and `firstUppercased` in the String extension.
+- Added `enumeratedArray` as an extension to `Collection`.
 
 ### v3.0.1